sound: core: Fix argument type errors in control.c
Fixed type mismatch errors in sound/core/control.c where a 'struct snd_ctl_elem_value' was being passed as an argument to functions that expect a pointer to 'struct snd_ctl_elem_value'. Corrected the calls to pass the address of the 'data' structure using the '&' operator.

Fix details:
- Changed 'err = snd_ctl_elem_read(card, data);' to 'err = snd_ctl_elem_read(card, &data);'
- Changed 'err = snd_ctl_elem_write(card, file, data);' to 'err = snd_ctl_elem_write(card, file, &data);'

Signed-off-by: Ruri <barryofc11@gmail.com>