drivers: input: fingerprint: Improve IRQ performance affinity
In this commit, we optimize the interrupt request flags (irqf) for
the FPC1020 and GF3208 fingerprint sensor drivers, as well as some
related touchscreen drivers. The change replaces the IRQF_PERF_AFFINE
flag with IRQF_PERF_CRITICAL to enhance the performance affinity of
interrupts.

For FPC1020:
- The irqf setting for FPC1020 now includes IRQF_TRIGGER_RISING,
  IRQF_ONESHOT, and IRQF_PERF_CRITICAL.
- If the "fpc,enable-wakeup" property is detected, IRQF_NO_SUSPEND
  is also added for handling wakeups.

For GF3208:
- The irqf setting for GF3208 now includes IRQF_TRIGGER_RISING,
  IRQF_ONESHOT, and IRQF_PERF_CRITICAL.

For Touchscreen drivers:
- We removed the ksu_input_hook condition as it is no longer needed.
- The deprecated IRQF_PERF_AFFINE flag is replaced with IRQF_PERF_CRITICAL
  in the irq setup.

These changes aim to enhance the real-time capabilities and performance of
interrupt handling in the specified drivers.

Signed-off-by: Akari <moenyanuwu@gmail.com>